Version 2.02 - implimented a function for user to set into different modes via LCD screen and 2 buttons.
Implimented 2 main concepts (state and mode) in order to solve this problem.
there are 2 states which are run state and set state.
1. runstate
Run state is when the device is running the functions it is made to do such as refilling the water, sensing water level, standing by for user to press button and etc.
The run states runs the modes that were selected by the user.
2. set state
Set state is when the device is running the normal functions but in a state where user can set into different modes.

There are 3 kinds of modes
1. auto mode - always sensing water level and refilling if water is low
2. reserve mode - only senses and refills on set time has been reached and when user has pressed the left button
3. manual mode - only senses and refills when user has pressed the left button

when device turns on, it goes in to usn state
During run state, device can switch into set state by pressing the right button.
In the set state, you can navigate to other modes by pressing the left button, and the right button to select into a mode, which will enter in to run state and run the selected mode.

in set mode, if user does not select a mode for more than 30 loop functions, device will automatically enter run state and run the prior mode that was running.

the time cannot be set by user but by changing the code currently. it is set to 9am by default.