Version 2.02.03 - Fixed issue 2.02.01
Interrupt bouncing issue
Fixed this issue by adding a boolean variable called interruptStandBy.
This variable is set to true when interrupt is triggered,
When this variable is true, the ISR functions does not run anything.
The variable is switched to false when a new loop restarts.
This structure prohibits the interrupt to trigger more than once in a row causeing bouncing issues.

Also made minor changes to the code for optimizition.